requirements.txt
boto3==1.28.14
numpy==1.25.1
open3d==0.17.0
pandas==2.0.3
scikit_learn==1.3.0
scipy==1.11.2
segments_ai==1.0.26
